Fixes bug 468 . Now data rows which contain an ON DELETE CASCADE type reference to a missing row will be deleted during the update.database process
--HG--
extra : convert_revision : svn%3A6fd5776a-9540-455a-916b-3609c81e8f66/trunk%4073

+++ src/org/apache/ddlutils/io/DataToDatabaseSink.java
@@ -80,6 +80,8 @@ public class DataToDatabaseSink implements DataSink
 //    private HashSet _mergedTables = new HashSet();
     /** Stores the database filter into the database. */
     private DatabaseFilter _databasefilter = null;
+    
+    private boolean _deleteInvalidRows=false;
 
     /**
      * Creates a new sink instance.
@@ -131,6 +133,10 @@ public class DataToDatabaseSink implements DataSink
         return _haltOnErrors;
     }
 
+    public void setDeleteInvalidRows(boolean deleteInvalidRows)
+    {
+    	_deleteInvalidRows=deleteInvalidRows;
+    }
     /**
      * Specifies whether this sink halts when an error happens during the insertion of a bean
      * into the database.
@@ -232,8 +238,8 @@ public class DataToDatabaseSink implements DataSink
         purgeBatchQueue();
         try
         {
-
-            _platform.deleteInvalidConstraintRows(_model, _haltOnErrors);
+        	if(_deleteInvalidRows)
+        		_platform.deleteInvalidConstraintRows(_model, _haltOnErrors);
             _platform.enableAllTriggers(_connection, _model, _haltOnErrors);
             _platform.enableAllFK(_connection, _model, _haltOnErrors);
             _connection.close();
